back/back_django/requirements.txt
django==3.2
djangorestframework
django-cors-headers
mysqlclient
Pillow
django-paypal
cloudinary
python-decouple